+++ bin/migrate-oats-data/documents/__init__.py
@@ -1,2 +1,4 @@
 from .documents import *
 from .application_documents import *
+from .documents_noi import *
+from .noi_documents import *
