+++ bin/migrate-oats-data/documents/documents_noi.py
@@ -1,7 +1,10 @@
 from db import inject_conn_pool
 
 """
-    This script links data from ALCS documents table to ALCS noi_documents table based on data from OATS.
+    This script connects to postgress version of OATS DB and transfers data from OATS documents table to ALCS document table.
+
+    NOTE:
+    Before performing document import you need to import applications from oats.
 """
 
 
@@ -11,34 +14,30 @@ def compile_document_insert_query(number_of_rows_to_insert):
     """
     documents_to_insert = ",".join(["%s"] * number_of_rows_to_insert)
     return f"""
-       insert into alcs.noi_document
-            (	
-                application_uuid ,
-                document_uuid ,
-                type_code ,
-                visibility_flags,
-                oats_document_id,
-                oats_application_id,
-                audit_created_by
-            )
+        INSERT INTO alcs."document" (oats_document_id, file_name, oats_application_id, "source", 
+                                    audit_created_by, file_key, mime_type, tags, "system") 
         VALUES {documents_to_insert} 
-        ON CONFLICT (oats_document_id, oats_application_id) DO UPDATE SET 
-            application_uuid = EXCLUDED.application_uuid, 
-            document_uuid = EXCLUDED.document_uuid, 
-            type_code = EXCLUDED.type_code,
-            visibility_flags = EXCLUDED.visibility_flags,
-            audit_created_by = EXCLUDED.audit_created_by;
+        ON CONFLICT (oats_document_id) DO UPDATE SET 
+            oats_document_id = EXCLUDED.oats_document_id, 
+            file_name = EXCLUDED.file_name, 
+            oats_application_id = EXCLUDED.oats_application_id, 
+            "source" = EXCLUDED."source", 
+            audit_created_by = EXCLUDED.audit_created_by, 
+            file_key = EXCLUDED.file_key, 
+            mime_type = EXCLUDED.mime_type,
+            tags = EXCLUDED.tags,
+            "system" = EXCLUDED."system"
     """
 
 
 @inject_conn_pool
-def process_noi_documents(conn=None, batch_size=10000):
+def process_documents_noi(conn=None, batch_size=10000):
     """
     function uses a decorator pattern @inject_conn_pool to inject a database connection pool to the function. It fetches the total count of documents and prints it to the console. Then, it fetches the documents to insert in batches using document IDs, constructs an insert query, and processes them.
     """
     with conn.cursor() as cursor:
         with open(
-            "noi/documents/noi_documents_total_count.sql", "r", encoding="utf-8"
+            "sql/documents_noi/documents_noi_total_count.sql", "r", encoding="utf-8"
         ) as sql_file:
             count_query = sql_file.read()
             cursor.execute(count_query)
@@ -49,13 +48,11 @@ def process_noi_documents(conn=None, batch_size=10000):
         successful_inserts_count = 0
         last_document_id = 0
 
-        with open(
-            "noi/documents/noi_documents.sql", "r", encoding="utf-8"
-        ) as sql_file:
+        with open("sql/documents_noi/documents_noi.sql", "r", encoding="utf-8") as sql_file:
             documents_to_insert_sql = sql_file.read()
             while True:
                 cursor.execute(
-                    f"{documents_to_insert_sql} WHERE oats_document_id > {last_document_id} ORDER BY oats_document_id"
+                    f"{documents_to_insert_sql} WHERE document_id > {last_document_id} ORDER BY document_id"
                 )
                 rows = cursor.fetchmany(batch_size)
                 if not rows:
@@ -66,17 +63,16 @@ def process_noi_documents(conn=None, batch_size=10000):
                     insert_query = compile_document_insert_query(
                         documents_to_be_inserted_count
                     )
-
                     cursor.execute(insert_query, rows)
                     conn.commit()
 
-                    last_document_id = rows[-1][4]
+                    last_document_id = rows[-1][0]
                     successful_inserts_count = (
                         successful_inserts_count + documents_to_be_inserted_count
                     )
 
                     print(
-                        f"retrieved/inserted items count: {documents_to_be_inserted_count}; total successfully inserted/updated items so far {successful_inserts_count}; last inserted oats_document_id: {last_document_id}"
+                        f"retrieved/inserted items count: {documents_to_be_inserted_count}; total successfully inserted/updated documents so far {successful_inserts_count}; last inserted oats_document_id: {last_document_id}"
                     )
                 except Exception as e:
                     conn.rollback()
@@ -88,12 +84,3 @@ def process_noi_documents(conn=None, batch_size=10000):
     print("Total amount of failed inserts:", failed_inserts)
 
 
-@inject_conn_pool
-def clean_noi_documents(conn=None):
-    print("Start noi documents cleaning")
-    with conn.cursor() as cursor:
-        cursor.execute(
-            "DELETE FROM alcs.noi_document WHERE audit_created_by = 'oats_etl';"
-        )
-        conn.commit()
-        print(f"Deleted items count = {cursor.rowcount}")
