+++ bin/migrate-oats-data/sql/documents_noi/documents_noi.sql
@@ -10,7 +10,7 @@
 		ON  osp.subject_property_id  = od.subject_property_id 
 			AND osp.alr_application_id = od.alr_application_id 
 	WHERE od.alr_application_id IS NOT NULL 
-       	  AND document_code IS NOT NULL
+       	  AND document_code IS NOT NULL 
 		  AND od.issue_id IS NULL 
 		  AND od.planning_review_id IS NULL	
 )
